[LIFT-18] Single to FirstOrDefault + Business email domain. Preparing for business domain tests

+++ Backend/Core/Business/Backend.Business-ExerciseType/ExerciseTypeRequests/Delete/DeleteExerciseTypeRequestHandler.cs
@@ -1,11 +1,11 @@
-﻿using System;
-using System.Threading;
-using System.Threading.Tasks;
-using Backend.Domain;
+﻿using Backend.Domain;
 using Backend.Domain.Entities.Exercises;
 using Backend.Infrastructure.Exceptions;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
+using System;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace Backend.Business.Exercises.ExerciseTypeRequests.Delete
 {
@@ -23,7 +23,10 @@ namespace Backend.Business.Exercises.ExerciseTypeRequests.Delete
             try
             {
                 var exerciseType =
-                    await _context.Tags.SingleAsync(x => x.Id == request.Id, cancellationToken);
+                    await _context.Tags.FirstOrDefaultAsync(x => x.Id == request.Id, cancellationToken);
+
+                if (exerciseType == null)
+                    throw new NotFoundException(nameof(ExerciseType), request.Id);
 
                 _context.Tags.Remove(exerciseType);
